github: Detect secondary rate limit
This commit adds logic to the git call retrier to detect the GitHub
secondary rate limit. When detecting this limit, the github client
will wait for the set time (1 min) plus a few seconds more.

The idea behind this is that when doing calls in parallel we don't
want to set loose all workers to the API at the same time.

Signed-off-by: Adolfo García Veytia (Puerco) <adolfo.garcia@uservers.net>