Add TokenFromProviders method
Since we are importing the CLI, we need to ensure we test for a
OIDC identity token before attempting to keyless sign. If not,
cosign will attempt to to continue the interacitive flow asking
the token from the user.

This commit adds a function which will return the token from the
options if set, and if not will query the cosign identity providers
to get it.

Signed-off-by: Adolfo García Veytia (Puerco) <adolfo.garcia@uservers.net>