Remove OS specific way to construct part of URL
We have reports on upstream dependency krel not being reliable on Windows.
https://github.com/kubernetes/release/issues/4110

In the referenced slack conversation krel is running directly on Windows
and error is 'repository not found'

When we construct git repo URL we used OS specific way of concatenating
part of the URL (filepath.Join) As a result we were getting backslash in
the URL (https://github.com/kubernetes\kubernetes), which can lead to
errors.